+++ src/exchange.ts
@@ -1,29 +1,18 @@
-import { Api, JsonRpc } from "eosjs"
-import { JsSignatureProvider } from "eosjs/dist/eosjs-jssig"
-import Fetch from "node-fetch"
+import { IConnection } from "./interfaces/connection"
 import { IExchangePair, IExchangeSettings, IExchangeToken } from "./interfaces/exchange"
 import { ITrxParamsArgument } from "./interfaces/transaction"
-import { amountToAssetString, getTokenAccount, setTransactionParams } from "./utils"
+import { setTransactionParams } from "./utils"
 
 export class ExchangeContract {
-    private readonly contractName: string = "equiexchange"
-    public readonly rpc: JsonRpc
-    public readonly api: Api
+    protected readonly contractName: string = "eos2dtdotcom"
+    public readonly connection: IConnection
 
-    constructor(nodeAddress: string, privateKeys: string[]) {
-        const fetch: any = Fetch // Workaroung to avoid incompatibility of fetch types in 'eosjs' and 'node-fetch'
-        this.rpc = new JsonRpc(nodeAddress, { fetch })
-        const signatureProvider = new JsSignatureProvider(privateKeys)
-        this.api = new Api({
-            rpc: this.rpc,
-            signatureProvider,
-            textDecoder: new TextDecoder(),
-            textEncoder: new TextEncoder()
-        })
+    constructor(connection: IConnection) {
+        this.connection = connection
     }
 
-    public async getSettings(): Promise<IExchangeSettings> {
-        const table = await this.rpc.get_table_rows({
+    public getSettings = async (): Promise<IExchangeSettings> => {
+        const table = await this.connection.rpc.get_table_rows({
             code: this.contractName,
             scope: this.contractName,
             table: "xchsettings"
@@ -31,8 +20,8 @@ export class ExchangeContract {
         return table.rows[0]
     }
 
-    public async getAllPairs(): Promise<IExchangePair[]> {
-        const table = await this.rpc.get_table_rows({
+    public getAllPairs = async (): Promise<IExchangePair[]> => {
+        const table = await this.connection.rpc.get_table_rows({
             code: this.contractName,
             scope: this.contractName,
             table: "xchpairs",
@@ -41,8 +30,8 @@ export class ExchangeContract {
         return table.rows
     }
 
-    public async getAllTokens(): Promise<IExchangeToken[]> {
-        const table = await this.rpc.get_table_rows({
+    public getAllTokens = async (): Promise<IExchangeToken[]> => {
+        const table = await this.connection.rpc.get_table_rows({
             code: this.contractName,
             scope: this.contractName,
             table: "xchtokens",
@@ -51,15 +40,17 @@ export class ExchangeContract {
         return table.rows
     }
 
-    public async getToken(tokenSymbol: string): Promise<IExchangeToken | undefined> {
+    public getToken = async (
+        tokenSymbol: string
+    ): Promise<IExchangeToken | undefined> => {
         const tokens = await this.getAllTokens()
         return tokens.find(token => token.token_symbol.split(",")[1] === tokenSymbol)
     }
 
-    public async getPair(
+    public getPair = async (
         fromCurrency: string,
         toCurrency: string
-    ): Promise<IExchangePair | undefined> {
+    ): Promise<IExchangePair | undefined> => {
         const allPairs = await this.getAllPairs()
         return allPairs.find(pair => {
             const base = pair.base_currency.split(",")[1]
@@ -71,31 +62,34 @@ export class ExchangeContract {
         })
     }
 
-    public async exchange(
+    public exchange = async (
         sender: string,
         fromCurrency: string,
         toCurrency: string,
         amount: number,
         transactionParams?: ITrxParamsArgument
-    ): Promise<any> {
+    ): Promise<any> => {
         const pair = await this.getPair(fromCurrency, toCurrency)
         if (!pair) throw new Error(`Pair ${fromCurrency}/${toCurrency} does not exist`)
 
         const trxParams = setTransactionParams(transactionParams)
         const authorization = [{ permission: trxParams.permission, actor: sender }]
-        const type = pair.base_currency.split(",")[1] === fromCurrency ? "sell" : "buy"
-        const receipt = await this.api.transact(
+
+        const receipt = await this.connection.api.transact(
             {
                 actions: [
                     {
-                        account: getTokenAccount(fromCurrency),
+                        account: await this.getTokenAccount(fromCurrency),
                         name: "transfer",
                         authorization,
                         data: {
                             from: sender,
                             to: this.contractName,
-                            quantity: amountToAssetString(amount, fromCurrency),
-                            memo: `{"pair_id": ${pair.pair_id}, "type": "${type}"}`
+                            quantity: await this.amountToAssetString(
+                                amount,
+                                fromCurrency
+                            ),
+                            memo: `{"pair_id": ${pair.pair_id}}`
                         }
                     }
                 ]
@@ -108,47 +102,77 @@ export class ExchangeContract {
         return receipt
     }
 
-    public async getExchangeRate(
+    public getExchangeRate = async (
         fromCurrency: string,
         toCurrency: string
-    ): Promise<number | undefined> {
+    ): Promise<number | undefined> => {
         const pair = await this.getPair(fromCurrency, toCurrency)
         if (!pair) throw new Error(`Pair ${fromCurrency}/${toCurrency} does not exist`)
 
-        const oracleName = "eosdtorclize"
+        let rate: number
 
-        const rates = (
-            await this.rpc.get_table_rows({
-                code: oracleName,
-                scope: oracleName,
-                table: "orarates",
-                limit: 100
-            })
-        ).rows
+        if (pair.price_type == 2) {
+            rate =
+                pair.base_currency == fromCurrency
+                    ? 1.0 / parseFloat(pair.price)
+                    : parseFloat(pair.price)
+        } else {
+            const oracleName = "eosdtorclize"
 
-        let ratesMap = new Map<string, number>()
+            const rates = (
+                await this.connection.rpc.get_table_rows({
+                    code: oracleName,
+                    scope: oracleName,
+                    table: "orarates",
+                    limit: 100
+                })
+            ).rows
 
-        ratesMap.set("EOS", 1)
-        for (const rate of rates) {
-            const symbol = rate.rate.match(/[A-Z]+/g)![0]
-            const value = parseFloat(rate.rate)
-            ratesMap.set(symbol, value)
+            const ratesMap = new Map<string, number>()
+
+            ratesMap.set("EOS", 1)
+            for (const rateEntry of rates) {
+                const symbol = rateEntry.rate.match(/[A-Z]+/g)![0]
+                const value = parseFloat(rateEntry.rate)
+                ratesMap.set(symbol, value)
+            }
+
+            // EOSDT rate is equal to USD
+            const fromKey = fromCurrency === "EOSDT" ? "USD" : fromCurrency
+            const toKey = toCurrency === "EOSDT" ? "USD" : toCurrency
+
+            if (!(ratesMap.has(fromKey) && ratesMap.has(toKey))) {
+                return undefined
+            }
+
+            rate = ratesMap.get(toKey)! / ratesMap.get(fromKey)!
         }
 
-        const fromKey = fromCurrency === "EOSDT" ? "USD" : fromCurrency
-        const toKey = toCurrency === "EOSDT" ? "USD" : toCurrency
+        const exchangeRate =
+            rate *
+            (pair.base_currency == fromCurrency
+                ? 1.0 / (1.0 + Number(pair.buy_slippage))
+                : 1.0 - Number(pair.sell_slippage))
 
-        if (!(ratesMap.has(fromKey) && ratesMap.has(toKey))) {
-            return undefined
-        }
+        return exchangeRate
+    }
 
-        let rate = ratesMap.get(fromKey)! / ratesMap.get(toKey)!
+    public getTokenAccount = async (assetSymbol: string): Promise<string> => {
+        const token = await this.getToken(assetSymbol)
+        if (token === undefined) throw new Error(`Unknown token symbol ${assetSymbol}`)
+        return token.token_account
+    }
 
-        rate *=
-            pair.base_currency == fromCurrency
-                ? 1 - pair.sell_slippage
-                : 1.0 / (1 + pair.buy_slippage)
+    public amountToAssetString = async (
+        amount: number | string,
+        assetSymbol: string
+    ): Promise<string> => {
+        const token = await this.getToken(assetSymbol)
+        if (token === undefined) throw new Error(`Unknown token symbol ${assetSymbol}`)
+        const decimals = Number(token.token_symbol.split(",")[0])
+        if (typeof amount === "string") amount = parseFloat(amount)
+        assetSymbol = assetSymbol.toUpperCase()
 
-        return rate
+        return `${amount.toFixed(decimals)} ${assetSymbol}`
     }
 }