+++ src/interfaces/connection.ts
@@ -1,0 +1,6 @@
+import { Api, JsonRpc } from "eosjs"
+
+export interface IConnection {
+    rpc: JsonRpc
+    api: Api
+}