+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
 
 All notable changes will be documented in this file.
+
+#### [1.0.2] - 2019-12-05
+
+Changes in accordance with changes in the exchange contract:
+
+-   Added manager account and price to IExchangePair interface